requirements.txt
Django==3.2.5
django-debug-toolbar==3.2.2